Rename martin-mbtiles crate to mbtiles (#976)
This simplifies usage as both a crate and as a tool
