+++ martin/Cargo.toml
@@ -66,8 +66,8 @@ futures.workspace = true
 itertools.workspace = true
 json-patch.workspace = true
 log.workspace = true
-martin-mbtiles.workspace = true
 martin-tile-utils.workspace = true
+mbtiles.workspace = true
 num_cpus.workspace = true
 pbf_font_tools.workspace = true
 pmtiles.workspace = true
