+++ martin/src/mbtiles/mod.rs
@@ -5,8 +5,8 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use log::trace;
-use martin_mbtiles::MbtilesPool;
 use martin_tile_utils::TileInfo;
+use mbtiles::MbtilesPool;
 use tilejson::TileJSON;
 
 use crate::file_config::FileError;
