+++ mbtiles/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "martin-mbtiles"
-version = "0.6.0"
+name = "mbtiles"
+version = "0.7.0"
 authors = ["Yuri Astrakhan <YuriAstrakhan@gmail.com>", "MapLibre contributors"]
 description = "A simple low-level MbTiles access and processing library, with some tile format detection and other relevant heuristics."
 keywords = ["mbtiles", "maps", "tiles", "mvt", "tilejson"]
