+++ mbtiles/src/errors.rs
